feat (campaign-edit): move posting start and end date fields to general info section
