+++ src/sections/campaign/manage/details/UpdateAdditionalOne.jsx
@@ -1,4 +1,3 @@
-import dayjs from 'dayjs';
 import * as Yup from 'yup';
 import PropTypes from 'prop-types';
 import { useForm } from 'react-hook-form';
@@ -7,13 +6,10 @@ import { yupResolver } from '@hookform/resolvers/yup';
 import { useMemo, useEffect, useCallback } from 'react';
 
 import { LoadingButton } from '@mui/lab';
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { DatePicker , LocalizationProvider } from '@mui/x-date-pickers';
 import { Box, Grid, Stack, FormLabel, Typography } from '@mui/material';
 
 import axiosInstance, { endpoints } from 'src/utils/axios';
 
-import Iconify from 'src/components/iconify';
 import { RHFUpload } from 'src/components/hook-form/rhf-upload';
 import FormProvider from 'src/components/hook-form/form-provider';
 import { RHFTextField, RHFMultiSelect } from 'src/components/hook-form';
@@ -69,8 +65,6 @@ const CONTENT_FORMAT_OPTIONS = [
 const UpdateAdditionalOneSchema = Yup.object().shape({
   socialMediaPlatform: Yup.array(),
   contentFormat: Yup.array(),
-  postingStartDate: Yup.mixed().nullable(),
-  postingEndDate: Yup.mixed().nullable(),
   mainMessage: Yup.string().nullable(),
   keyPoints: Yup.string().nullable(),
   toneAndStyle: Yup.string().nullable(),
@@ -88,12 +82,6 @@ const UpdateAdditionalOne = ({ campaign, campaignMutate }) => {
     () => ({
       socialMediaPlatform: campaign?.campaignBrief?.socialMediaPlatform || [],
       contentFormat: additionalDetails?.contentFormat || [],
-      postingStartDate: campaign?.campaignBrief?.postingStartDate
-        ? new Date(campaign.campaignBrief.postingStartDate).toISOString()
-        : null,
-      postingEndDate: campaign?.campaignBrief?.postingEndDate
-        ? new Date(campaign.campaignBrief.postingEndDate).toISOString()
-        : null,
       mainMessage: additionalDetails?.mainMessage || '',
       keyPoints: additionalDetails?.keyPoints || '',
       toneAndStyle: additionalDetails?.toneAndStyle || '',
@@ -128,12 +116,6 @@ const UpdateAdditionalOne = ({ campaign, campaignMutate }) => {
     }
   }, [campaign, defaultValues, reset]);
 
-  const postingStartDate = watch('postingStartDate');
-  const postingEndDate = watch('postingEndDate');
-
-  const postingStartDateDayjs = postingStartDate ? dayjs(postingStartDate) : null;
-  const postingEndDateDayjs = postingEndDate ? dayjs(postingEndDate) : null;
-
   const onSubmit = useCallback(
     async (data) => {
       try {
@@ -143,8 +125,6 @@ const UpdateAdditionalOne = ({ campaign, campaignMutate }) => {
         // Additional Details 1 fields
         formData.append('socialMediaPlatform', JSON.stringify(data.socialMediaPlatform || []));
         formData.append('contentFormat', JSON.stringify(data.contentFormat || []));
-        formData.append('postingStartDate', data.postingStartDate || '');
-        formData.append('postingEndDate', data.postingEndDate || '');
         formData.append('mainMessage', data.mainMessage || '');
         formData.append('keyPoints', data.keyPoints || '');
         formData.append('toneAndStyle', data.toneAndStyle || '');
@@ -249,78 +229,6 @@ const UpdateAdditionalOne = ({ campaign, campaignMutate }) => {
                 options={CONTENT_FORMAT_OPTIONS}
               />
             </FormField>
-
-            {/* Campaign Posting Period */}
-            <Typography variant="subtitle2" fontWeight={600}>
-              Campaign Posting Period
-            </Typography>
-            <LocalizationProvider dateAdapter={AdapterDayjs}>
-              <Grid container spacing={2}>
-                <Grid item xs={12} sm={6}>
-                  <FormField>
-                    <DatePicker
-                      defaultValue={postingStartDateDayjs}
-                      format="DD/MM/YY"
-                      onChange={(newValue) => {
-                        setValue('postingStartDate', newValue ? newValue.toISOString() : null, {
-                          shouldValidate: true,
-                          shouldDirty: true,
-                        });
-                        // Auto-set postingEndDate to +5 days if not manually set
-                        if (newValue) {
-                          const newEndDate = dayjs(newValue).add(5, 'day');
-                          // Only auto-set if postingEndDate is empty or was previously auto-set
-                          if (!postingEndDateDayjs || postingEndDateDayjs.isSame(dayjs(postingStartDate).add(5, 'day'))) {
-                            setValue('postingEndDate', newEndDate.toISOString(), {
-                              shouldValidate: true,
-                              shouldDirty: true,
-                            });
-                          }
-                        }
-                      }}
-                      slots={{
-                        openPickerIcon: () => (
-                          <Iconify icon="meteor-icons:calendar" width={22} />
-                        ),
-                      }}
-                      slotProps={{
-                        textField: {
-                          fullWidth: true,
-                          placeholder: 'Start Date',
-                          error: false,
-                        },
-                      }}
-                    />
-                  </FormField>
-                </Grid>
-                <Grid item xs={12} sm={6}>
-                  <FormField>
-                    <DatePicker
-                      value={postingEndDateDayjs}
-                      onChange={(newValue) => {
-                        setValue('postingEndDate', newValue ? newValue.toISOString() : null, {
-                          shouldValidate: true,
-                          shouldDirty: true,
-                        });
-                      }}
-                      format="DD/MM/YY"
-                      slots={{
-                        openPickerIcon: () => (
-                          <Iconify icon="meteor-icons:calendar" width={22} />
-                        ),
-                      }}
-                      slotProps={{
-                        textField: {
-                          fullWidth: true,
-                          placeholder: 'End Date',
-                          error: false,
-                        },
-                      }}
-                    />
-                  </FormField>
-                </Grid>
-              </Grid>
-            </LocalizationProvider>
 
             {/* Main Message/Theme */}
             <FormField label="Main Message/Theme - What's the core message?">
